backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
google-genai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-dotenv>=1.0.0
